+++ samples/sampler/stories/qa/WizardPanels.js
@@ -1,3 +1,4 @@
+/* eslint-disable react/jsx-no-bind */
 import {action} from '@enact/storybook-utils/addons/actions';
 import {boolean, number, select} from '@enact/storybook-utils/addons/knobs';
 import React from 'react';
@@ -8,6 +9,7 @@ import Button from '@enact/sandstone/Button';
 import Icon from '@enact/sandstone/Icon';
 import Item from '@enact/sandstone/Item';
 import {Scroller} from '@enact/sandstone/Scroller';
+import {Panel} from '@enact/sandstone/WizardPanels';
 import WizardPanels, {WizardPanelsBase} from '@enact/sandstone/WizardPanels';
 
 import {mergeComponentMetadata} from '@enact/storybook-utils';
@@ -27,6 +29,33 @@ const inputData = {
 };
 
 WizardPanels.displayName = 'WizardPanels';
+
+class WizardPanelsWidthFooterButtons extends React.Component {
+	constructor () {
+		super();
+		this.state = {
+			index: 0
+		};
+	}
+
+	render () {
+		return (
+			<WizardPanels noAnimation index={this.state.index}>
+				<Panel title={'Panel0'} subtitle={'subtitle'} nextButton={<Button>Next</Button>}>
+					<footer>
+						<Button>Dummy</Button>
+						<Button onClick={() => this.setState({index: 1})}>Next</Button>
+					</footer>
+				</Panel>
+				<Panel title={'Panel1'} subtitle={'subtitle'}>
+					<footer>
+						<Button onClick={() => this.setState({index: 0})}>Previous</Button>
+					</footer>
+				</Panel>
+			</WizardPanels>
+		);
+	}
+}
 
 storiesOf('WizardPanels', module)
 	.add(
@@ -111,4 +140,9 @@ storiesOf('WizardPanels', module)
 				noPanel: true
 			}
 		}
+	).add(
+		'with footter buttons',
+		() => (
+			<WizardPanelsWidthFooterButtons />
+		)
 	);